Fix formatting, alignment. repeated renders.
